requirements.txt
dash==2.17.1
pandas==2.2.3
plotly-express==0.4.1


